refactor: Creating a folder for the profile domain and moving components that are strongly related to it
